Валидация формы ref
